# docs/splunky-api-contract.yaml
openapi: 3.0.3
info:
  title: Splunky API Contract
  version: 0.2.0
  description: |
    Contract aligned to the current Splunky implementation.

    Current frontend MVP is summary-only:
    - Splunk-backed login
    - investigation execution
    - AI summary
    - raw log preview
    - executed SPL

    Raw Splunk logs returned by the API are transient previews and must not be
    persisted as durable database records.

servers:
  - url: http://localhost:8080
    description: Local backend

tags:
  - name: Session
  - name: Investigation
  - name: Query Template
  - name: LLM Credential

paths:
  /api/sessions/splunk-login:
    post:
      tags: [Session]
      summary: Start a Splunk-backed session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplunkLoginRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/sessions/current:
    get:
      tags: [Session]
      summary: Get the current session
      security:
        - splunkySessionId: []
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Session]
      summary: Logout the current session
      security:
        - splunkySessionId: []
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/llm-credential/status:
    get:
      tags: [LLM Credential]
      summary: Check current LLM credential status
      responses:
        '200':
          description: Current provider credential status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmCredentialStatus'

  /api/query-templates:
    get:
      tags: [Query Template]
      summary: List enabled query templates
      responses:
        '200':
          description: Enabled templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QueryTemplateRecord'

  /api/investigations:
    post:
      tags: [Investigation]
      summary: Start an investigation
      security:
        - splunkySessionId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartInvestigationRequest'
      responses:
        '200':
          description: Investigation created and active result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investigation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/investigations/{investigationId}:
    get:
      tags: [Investigation]
      summary: Get an investigation
      security:
        - splunkySessionId: []
      parameters:
        - name: investigationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Investigation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investigation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/investigations/{investigationId}/follow-ups:
    post:
      tags: [Investigation]
      summary: Submit a follow-up
      description: |
        Backend support exists, but the current frontend MVP does not surface the
        assistant UI. This endpoint is retained for future use and testing.
      security:
        - splunkySessionId: []
      parameters:
        - name: investigationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowUpRequest'
      responses:
        '200':
          description: Follow-up handled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowUpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    splunkySessionId:
      type: apiKey
      in: header
      name: X-Splunky-Session-Id

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or expired session, or invalid Splunk credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time

    SplunkLoginRequest:
      type: object
      required: [splunkUsername, splunkPassword, environment]
      properties:
        splunkUsername:
          type: string
        splunkPassword:
          type: string
        environment:
          type: string
          enum: [DEV, PROD_ON_PREM, PROD_AWS]

    UserProfile:
      type: object
      required: [userId, username, displayName, status]
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string

    SessionResponse:
      type: object
      required:
        - sessionId
        - user
        - splunkUsername
        - environment
        - status
        - startedAt
        - expiresAt
      properties:
        sessionId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserProfile'
        splunkUsername:
          type: string
        environment:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
          nullable: true

    LlmCredentialStatus:
      type: object
      required:
        - configured
        - provider
        - status
        - apiKeyConfigured
        - sessionTokenConfigured
      properties:
        configured:
          type: boolean
        provider:
          type: string
        status:
          type: string
          enum: [NOT_CONFIGURED, ACTIVE, EXPIRED]
        apiKeyConfigured:
          type: boolean
        sessionTokenConfigured:
          type: boolean
        sessionTokenExpiresAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true

    StartInvestigationRequest:
      type: object
      required: [rawText]
      properties:
        rawText:
          type: string
        selectedInputTypes:
          type: array
          items:
            type: string
        timeRange:
          $ref: '#/components/schemas/TimeRange'
        apiName:
          type: string
          nullable: true

    FollowUpRequest:
      type: object
      properties:
        prompt:
          type: string
          nullable: true
        message:
          type: string
          nullable: true

    FollowUpResponse:
      type: object
      required: [investigation, actionType]
      properties:
        investigation:
          $ref: '#/components/schemas/Investigation'
        actionType:
          type: string
        requeryStatus:
          $ref: '#/components/schemas/RequeryStatus'
          nullable: true

    RequeryStatus:
      type: object
      required: [title, changes, queryType]
      properties:
        title:
          type: string
        changes:
          type: array
          items:
            type: string
        queryType:
          type: string

    Investigation:
      type: object
      required:
        - id
        - activeRunId
        - createdAt
        - updatedAt
        - input
        - runs
        - activeResult
        - conversation
      properties:
        id:
          type: string
        activeRunId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        input:
          $ref: '#/components/schemas/InvestigationInput'
        runs:
          type: array
          items:
            $ref: '#/components/schemas/InvestigationRunSummary'
        activeResult:
          $ref: '#/components/schemas/InvestigationResult'
        conversation:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'

    InvestigationInput:
      type: object
      required: [rawText, detectedTypes, timeRange]
      properties:
        rawText:
          type: string
        detectedTypes:
          type: array
          items:
            type: string
        timeRange:
          $ref: '#/components/schemas/TimeRange'
        apiName:
          type: string
          nullable: true
        correlationId:
          type: string
          nullable: true

    TimeRange:
      type: object
      required: [label, from, to, timezone]
      properties:
        label:
          type: string
        from:
          type: string
        to:
          type: string
        timezone:
          $ref: '#/components/schemas/TimezoneOption'

    TimezoneOption:
      type: object
      required: [label, offset]
      properties:
        label:
          type: string
        offset:
          type: string

    InvestigationRunSummary:
      type: object
      required: [runId, runNumber, title, action, summary, createdAt]
      properties:
        runId:
          type: string
        runNumber:
          type: integer
        title:
          type: string
        userQuestion:
          type: string
          nullable: true
        action:
          type: string
        summary:
          type: string
        createdAt:
          type: string

    InvestigationResult:
      type: object
      required:
        - investigationId
        - runId
        - runNumber
        - context
        - summary
        - timeline
        - serviceGraph
        - sequence
        - downstreamCalls
        - rawLogs
        - queries
        - suggestedFollowUps
      properties:
        investigationId:
          type: string
        runId:
          type: string
        runNumber:
          type: integer
        context:
          $ref: '#/components/schemas/InvestigationContext'
        summary:
          $ref: '#/components/schemas/DiagnosisSummary'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        serviceGraph:
          $ref: '#/components/schemas/ServiceGraph'
        sequence:
          $ref: '#/components/schemas/SequenceViewModel'
        downstreamCalls:
          type: array
          items:
            $ref: '#/components/schemas/DownstreamCall'
        rawLogs:
          type: array
          items:
            $ref: '#/components/schemas/RawLogEntry'
        queries:
          type: array
          items:
            $ref: '#/components/schemas/SplQueryRecord'
        suggestedFollowUps:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedFollowUp'

    InvestigationContext:
      type: object
      required: [timeRange, lastRunAt]
      properties:
        timeRange:
          $ref: '#/components/schemas/TimeRange'
        correlationId:
          type: string
          nullable: true
        apiName:
          type: string
          nullable: true
        lastRunAt:
          type: string

    DiagnosisSummary:
      type: object
      required:
        - status
        - confidence
        - logsFound
        - affectedServices
        - evidence
        - recommendedActions
      properties:
        status:
          type: string
        finalHttpStatus:
          type: integer
          nullable: true
        failurePoint:
          type: string
          nullable: true
        rootCauseHypothesis:
          type: string
          nullable: true
        confidence:
          type: string
        logsFound:
          type: integer
        affectedServices:
          type: array
          items:
            type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceItem'
        recommendedActions:
          type: array
          items:
            type: string

    EvidenceItem:
      type: object
      required: [id, title, description, relatedLogIds, relatedTimelineEventIds]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
          nullable: true
        service:
          type: string
          nullable: true
        relatedLogIds:
          type: array
          items:
            type: string
        relatedTimelineEventIds:
          type: array
          items:
            type: string

    TimelineEvent:
      type: object
      required: [id, timestamp, service, eventType, status, message, relatedLogIds]
      properties:
        id:
          type: string
        timestamp:
          type: string
        service:
          type: string
        eventType:
          type: string
        status:
          type: string
        durationMs:
          type: integer
          nullable: true
        message:
          type: string
        relatedLogIds:
          type: array
          items:
            type: string
        downstreamCallId:
          type: string
          nullable: true

    ServiceGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/ServiceNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/ServiceEdge'

    ServiceNode:
      type: object
      required: [id, serviceName, status, logCount, errorCount]
      properties:
        id:
          type: string
        serviceName:
          type: string
        platform:
          type: string
          nullable: true
        status:
          type: string
        logCount:
          type: integer
        errorCount:
          type: integer

    ServiceEdge:
      type: object
      required: [id, source, target, label, status, evidenceType, relatedLogIds]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        label:
          type: string
        operation:
          type: string
          nullable: true
        status:
          type: string
        latencyMs:
          type: integer
          nullable: true
        evidenceType:
          type: string
        relatedLogIds:
          type: array
          items:
            type: string
        downstreamCallId:
          type: string
          nullable: true

    SequenceViewModel:
      type: object
      required: [participants, messages]
      properties:
        participants:
          type: array
          items:
            $ref: '#/components/schemas/SequenceParticipant'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/SequenceMessage'

    SequenceParticipant:
      type: object
      required: [id, label, serviceName]
      properties:
        id:
          type: string
        label:
          type: string
        serviceName:
          type: string

    SequenceMessage:
      type: object
      required: [id, from, to, label, timestamp, status, relatedLogIds]
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        label:
          type: string
        timestamp:
          type: string
        status:
          type: string
        durationMs:
          type: integer
          nullable: true
        relatedLogIds:
          type: array
          items:
            type: string
        downstreamCallId:
          type: string
          nullable: true

    DownstreamCall:
      type: object
      required: [id, caller, downstream, operation, status, relatedLogIds]
      properties:
        id:
          type: string
        caller:
          type: string
        downstream:
          type: string
        operation:
          type: string
        endpoint:
          type: string
          nullable: true
        status:
          type: string
        latencyMs:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true
        relatedLogIds:
          type: array
          items:
            type: string

    RawLogEntry:
      type: object
      required: [id, timestamp, level, service, message, fields]
      properties:
        id:
          type: string
        timestamp:
          type: string
        level:
          type: string
        service:
          type: string
        eventType:
          type: string
          nullable: true
        message:
          type: string
        fields:
          type: object
          additionalProperties: true
        splunkUrl:
          type: string
          nullable: true

    SplQueryRecord:
      type: object
      required:
        - id
        - templateName
        - reason
        - spl
        - timeRange
        - resultCount
        - executionDurationMs
        - status
      properties:
        id:
          type: string
        templateName:
          type: string
        reason:
          type: string
        spl:
          type: string
        timeRange:
          $ref: '#/components/schemas/TimeRange'
        resultCount:
          type: integer
        executionDurationMs:
          type: integer
        status:
          type: string
        splunkUrl:
          type: string
          nullable: true

    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
        actionType:
          type: string
          nullable: true
        relatedRunId:
          type: string
          nullable: true

    SuggestedFollowUp:
      type: object
      required: [id, label, prompt, actionType]
      properties:
        id:
          type: string
        label:
          type: string
        prompt:
          type: string
        actionType:
          type: string

    QueryTemplateRecord:
      type: object
      required:
        - templateId
        - templateKey
        - name
        - templateSpl
        - defaultTimeWindowMinutes
        - maxTimeWindowMinutes
        - maxResultCount
        - enabled
      properties:
        templateId:
          type: string
          format: uuid
        templateKey:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        templateSpl:
          type: string
        defaultTimeWindowMinutes:
          type: integer
        maxTimeWindowMinutes:
          type: integer
        maxResultCount:
          type: integer
        enabled:
          type: boolean